last commit before massive changes
This version doesn't work (for example, it improperly calls a private
method of the parent class, _url_collapse_path_split), which doesn't
even exist in Python 2.7.5.  I'm committing it as a potential restore
point in case I *really* blow it in subsequent steps, as I aim to...
o  implement only the Stuy/Adobe setup
o  guess that Adobe identifies a cgi script by the first element of the
slash-separated url that ends with “.py”
o  mimic Adobe’s ability to handle extra-path, since it’s easy given the
other changes
o  maintain the methods defined in CGIHTTPHandler, in case someone wants
to expand the capabilities in the future.